riskmap-funcional/requirements.txt
pydantic
python-dotenv
PyYAML
Flask
Flask-Cors
pandas
numpy
requests
werkzeug
Jinja2
Flask-Babel
feedparser
APScheduler>=3.9.0
opencv-python
openai
transformers
groq
aiohttp
schedule
